requirements.txt
deepspeed>=0.14.0
transformers>=4.40.0
torch>=2.1.0
datasets>=2.18.0
accelerate>=0.29.0
peft>=0.10.0
trl>=0.8.0
scikit-learn>=1.3.0
jieba>=0.42.1
pyyaml>=6.0
requests>=2.31.0
